Implemented frontend with yew, can view all posts, and setup to mirror api structurs. Also configured the readme

+++ rust-micro-cms/src/database/authors.rs
@@ -1,4 +1,4 @@
-use crate::models::{NewAuthor,Author, UpdateAuthor, Authors};
+use crate::models::{Author, Authors, NewAuthor, UpdateAuthor};
 use anyhow::Result;
 use sqlx::sqlite::SqlitePool;
 
@@ -20,23 +20,21 @@ pub async fn insert_new_author(pool: &SqlitePool, author: &NewAuthor) -> Result<
     let inserted_author = sqlx::query_as!(
         NewAuthor,
         "INSERT INTO author (first_name,last_name) VALUES (?,?) RETURNING *",
-        author.first_name, author.last_name
+        author.first_name,
+        author.last_name
     )
-    .fetch_one(pool) 
+    .fetch_one(pool)
     .await?;
 
     Ok(inserted_author)
 }
 
 pub async fn fetch_all_authors(pool: &SqlitePool) -> Result<Authors> {
-    let authors = sqlx::query_as!(
-        Author,
-        "SELECT * FROM author WHERE deleted = FALSE"
-    )
-    .fetch_all(pool)
-    .await?;
-    
-    Ok(Authors{authors})
+    let authors = sqlx::query_as!(Author, "SELECT * FROM author WHERE deleted = FALSE")
+        .fetch_all(pool)
+        .await?;
+
+    Ok(Authors { authors })
 }
 
 pub async fn fetch_author_by_id(pool: &SqlitePool, author_id: i32) -> Result<Option<Author>> {
@@ -55,7 +53,9 @@ pub async fn update_author(pool: &SqlitePool, new_author: UpdateAuthor) -> Resul
     sqlx::query_as!(
         UpdateAuthor,
         "UPDATE author SET first_name = ?, last_name = ? WHERE id = ?",
-        new_author.first_name, new_author.last_name, new_author.id
+        new_author.first_name,
+        new_author.last_name,
+        new_author.id
     )
     .fetch_optional(pool)
     .await?;
@@ -64,31 +64,22 @@ pub async fn update_author(pool: &SqlitePool, new_author: UpdateAuthor) -> Resul
 
 //could cause conflict due to key restraints with posts
 pub async fn delete_author(pool: &SqlitePool, author_id: i32) -> Result<()> {
-    sqlx::query!(
-        "DELETE FROM author WHERE id = ?",
-        author_id
-    )
-    .execute(pool)
-    .await?;
+    sqlx::query!("DELETE FROM author WHERE id = ?", author_id)
+        .execute(pool)
+        .await?;
     Ok(())
 }
 
 pub async fn soft_delete_author(pool: &SqlitePool, author_id: i32) -> Result<()> {
-    sqlx::query!(
-        "UPDATE author SET deleted = TRUE WHERE id = ?",
-        author_id
-    )
-    .execute(pool)
-    .await?;
+    sqlx::query!("UPDATE author SET deleted = TRUE WHERE id = ?", author_id)
+        .execute(pool)
+        .await?;
     Ok(())
 }
 
 pub async fn reactivate_author(pool: &SqlitePool, author_id: i32) -> Result<()> {
-    sqlx::query!(
-        "UPDATE author SET deleted = FALSE WHERE id = ?",
-        author_id
-    )
-    .execute(pool)
-    .await?;
+    sqlx::query!("UPDATE author SET deleted = FALSE WHERE id = ?", author_id)
+        .execute(pool)
+        .await?;
     Ok(())
 }
